+++ MusicDiary/MusicDiary/DataStructures.swift
@@ -36,3 +36,10 @@ struct ContentData {
     
 }
 
+struct MusicStruct {
+    var musicTitle: String?
+    var musicArtist: String?
+    var musicCoverUrl: URL?
+    var musicLyrics: String?
+    var musicID: Int?
+}
