--- MusicDiary/MusicDiary/Yujin/MusicStructure.swift
@@ -1,15 +1,0 @@
-//
-//  SongStructure.swift
-//  MusicDiary
-//
-//  Created by 강유진 on 2021/02/04.
-//
-import UIKit
-
-struct MusicStruct {
-    var musicTitle: String?
-    var musicArtist: String?
-    var musicCoverUrl: URL?
-    var musicLyrics: String?
-    var musicID: Int?
-}
